feat: Add minting functionality for NFTs
This commit adds the functionality to mint NFTs in the application. It introduces a new function called `mintNft` that is triggered when the user clicks on the "Mint Your NFT!" button. The function creates a programmable transaction block using the `Transaction` class from the `@mysten/sui/transactions` package. It then calls the wallet to sign and execute the transaction block, minting the NFT. If successful, a success message is displayed, otherwise an error message is shown.

The commit also includes some additional imports and updates to the `Index` component, such as importing the `useWallet` hook from `@suiet/wallet-kit` and adding a conditional rendering for the "Mint Your NFT!" button based on the wallet's connection status.